.github/CC_PROJECT/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
python-dotenv==1.0.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-mgmt-web==3.0.0
gunicorn==21.2.0
email-validator==2.1.0.post1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
Werkzeug==3.0.1 